Made activity statements clearer
The biggest change was to the 6 systems activity to make it clearer what the outcome should be. Also reworded the statements under priority 2 so they start with a verb instead of 'we'. A few typos have been corrected and abbreviations written in full. 

These strategic priorities and activities have my consent.